% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fit.R
\name{add_fit}
\alias{add_fit}
\title{Add predicted values to a dataframe}
\usage{
add_fit(dat, mdl, pterm = NULL, sterm = NULL, se.mult = 2)
}
\arguments{
\item{dat}{A dataframe. It is usually a "new" dataframe for prediction, which usually has all the combination of predictors, created by expand.grid.}

\item{mdl}{A mgcv::gam/bam model. The response variable should be a normal distribution.}

\item{pterm}{A numeric vector to indicate which parametric terms to be used for prediction. This number can be known by base::summary (mgcv::summary.gam). If the parametric term shows up in the first position except the intercept term, this number is 1. If it is the second, the number is 2, and so on. However, ignore duplicated factor terms with different levels, to count this number. Please see the example below for details.}

\item{sterm}{A numeric vector to indicate which smooth terms to be used for prediction. This number is the same as where the smooth term shows up in mgcv::summary.gam.}
}
\value{
A dataframe, which has predicted values.
}
\description{
This function takes a dataframe and a GAM model and adds predicted values from the GAM model to the dataframe. The terms to be used for prediction can be set by the arguments "pterm" and "sterm" for parametric and smooth terms each. The current version of this function assumes the input model is a GAM model (mgcv::gam/bam) and the model was fitted with the response variable being a normal distribution.
}
\examples{
set.seed(10)
dat = gamSim(n=300)
set.seed(10)
dat$f = factor(sample(letters[1:3],300,replace=TRUE))
set.seed(10)
dat$g = factor(sample(LETTERS[1:3],300,replace=TRUE))
mdl = gam(y ~ f + g + s(x0, by=f) + te(x1, x2, by=g), data=dat)
vclist = list(f=list(), x0=list(edge=0, len.out=10),
	      x1=list(edge=0, len.out=10),
	      x2=list(edge=0, len.out=10))
ndat = create_ndat(mdl$model, vclist)
ndat = add_fit(ndat, mdl, pterm=2, sterm=5)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
