% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character_df.R
\name{as_character_df}
\alias{as_character_df}
\title{Selective conversion of factor to character columns}
\usage{
as_character_df(dat)
}
\arguments{
\item{dat}{A dataframe, some of whose columns you would like to convert to character columns.}
}
\value{
A new dataframe, where the columns which were originally factors are characters now.
}
\description{
Converts only factor columns into character columns.
}
\examples{
dat1 = data.frame(label=LETTERS[1:3], value=seq(10,30,10), other=letters[1:3])
dat2 = as_character_df(dat1)
str(dat1)
str(dat2)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
