% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{The R equivalent of the assert function in python}
\usage{
assert(bool)
}
\arguments{
\item{bool}{A boolean. It must be TRUE or FALSE.}
}
\value{
This function returns nothing.
}
\description{
If the only argument "x" is TRUE, this function does nothing. Otherwise, it gives a pseudo AssersionError.
}
\examples{
assert(TRUE)
assert(FALSE)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
