% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Curvature index for a given point}
\usage{
curvature(func, x)
}
\arguments{
\item{func}{A callable function, on which you want to get a curvature index. It has "x" as an argument.}

\item{x}{A numeric, whose length is 1. You get a curvature index on this point.}
}
\value{
A single value that represents how sharply a given function bends at a given point (i.e. a curvature index).
}
\description{
Calculates how pointy a given function is at a given point.
}
\examples{
func01 = function ( x ) 2*x^2
func02 = function ( x ) 0.5*x^2
print(curvature(func01, 0))
print(curvature(func02, 0))
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
