% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_outliers.R
\name{cutoff_outliers}
\alias{cutoff_outliers}
\title{Cutting off outliers}
\usage{
cutoff_outliers(vec, type = "sd", p = 0.1, value = FALSE)
}
\arguments{
\item{vec}{Variable as a vector, whose outliers to be cut off.}

\item{type}{Criterion to be used to define outliers.}

\item{p}{Quantile point. The values dropping out of the first p and the last p quantiles are removed.}

\item{value}{Return the remaining values if value=TRUE, else corresponding booleans are returned.}
}
\value{
Vector of booleans in the same length as the original vector if value=FALSE. If value=TRUE, outliers are removed and a new vector without the outliers is returned.
}
\description{
Cut off outliers by some different criteria. With type='sd', the values that lie in the outside of 2*sd(v) are removed. With type='quantile', the values are removed if they lie out of the quantiles defined by p.
}
\examples{
v = rchisq(n=1000, df=3)
v = list(v)
v[[2]] = cutoff_outliers(vec=v[[1]], type='sd', value=TRUE)
v[[3]] = cutoff_outliers(vec=v[[1]], type='quantile', value=TRUE)
par(mfrow=c(1,3))
for (i in v) {
    boxplot(i)
}
dev.off()
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
