% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_filename.R
\name{gen_filename}
\alias{gen_filename}
\title{Generating a file name}
\usage{
gen_filename(prefix = "", middle = c(), suffix = "", sep = "_")
}
\arguments{
\item{prefix}{A character string, which is attached to the beginning of the generated file name with '_' after it.}

\item{middle}{A character string or vector, which is inserted between prefix and suffix with '_' between each element if middle is a vector of strings.}

\item{suffix}{A character string, which is attached to the end of the generated file name with '.' before it.}

\item{sep}{A character string, which is used to concatenate all the elements from prefix, middle, and suffix. It is '_' as default.}
}
\value{
A concatenated string from the arguments.
}
\description{
Generates a file name with a prefix, a middle part, and a suffix.
}
\examples{
pfx = 'plot'
mid = c('foo', 'bar', 'qux')
sfx = 'png'
fn01 = file_name(prefix=pfx, middle=mid, suffix=sfx)
fn02 = file_name(prefix=pfx, middle=mid, suffix=sfx, sep='-')
print(fn01)
print(fn02)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
