% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_transform.R
\name{my_transform}
\alias{my_transform}
\title{Simple variable transformation}
\usage{
my_transform(vec, type = "log")
}
\arguments{
\item{vec}{Variable as a vector, which to be transformed.}

\item{type}{Which transformation method to use. Log- and BoxCox-transformations are implemented so far.}
}
\value{
Vector of transformed values.
}
\description{
Transform a variable by the BoxCox transformation or taking its log values.
}
\examples{
v = rchisq(n=1000, df=3)
v = list(v)
v[[2]] = my_transform(vec=v[[1]], type='log')
v[[3]] = my_transform(vec=v[[1]], type='boxcox')
names(v) = c('Original', 'Log', 'BoxCox')
par(mfrow=c(1,3))
for (i in seq_along(v)) {
    hist(v[[i]], main=names(v)[i], xlab='Values')
}
dev.off()
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
