% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm01.R
\name{norm01}
\alias{norm01}
\title{Normalization of a vector between 0 and 1}
\usage{
norm01(x)
}
\arguments{
\item{x}{A numeric vector, which you want to make into the range from 0 to 1.}
}
\value{
A new numeric vector, whose values are now normalized between 0 and 1.
}
\description{
This function normalizes a vector, so the whole range of the vector fits between 0 and 1.
}
\examples{
set.seed(0)
vec = rnorm(10)
vec = norm01(vec)
print(vec)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
