% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization of a vector into an arbitrary range}
\usage{
normalize(x, norm.range = c(0, 1))
}
\arguments{
\item{x}{A numeric vector, which are normalized.}

\item{norm.range}{A numeric vector, representing the range to which the vector (x) is normalized. Its length must be 2.}
}
\value{
A new numeric vector, whose values are now normalized to the new range defined by norm.range.
}
\description{
This function normalizes a vector of numerics to an arbitrary range. The range is \link{0,1} as default.
}
\examples{
set.seed(0)
vec00 = rnorm(100)
vec01 = normalize(vec00)
vec02 = normalize(vec00, norm.range=c(-0.5,0.5))
par(mfrow=c(1,3))
rng00 = round(range(vec00), 2)
rng01 = round(range(vec01), 2)
rng02 = round(range(vec02), 2)
ttl00 = paste(as.character(rng00),collapse=', ')
ttl01 = paste(as.character(rng01),collapse=', ')
ttl02 = paste(as.character(rng02),collapse=', ')
hist(vec00, main=sprintf('vec00, range=[\%s]', ttl00))
hist(vec01, main=sprintf('vec01, range=[\%s]', ttl01))
hist(vec02, main=sprintf('vec02, range=[\%s]', ttl02))
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
