% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Multiple value replacement in a vector}
\usage{
replace_values(
  vec,
  from = c("aː", "a", "iː", "ɪ", "uː", "ʊ"),
  to = c("a", "a", "i", "i", "u", "u")
)
}
\arguments{
\item{vec}{A vector to which you want to execute replacements.}

\item{from}{A vector of strings which you want to replace for something else.}

\item{to}{A vector of strings with the same length as 'from'. Target strings specified by 'from' are replaced by the corresponding values specified by this variable ('to').}
}
\value{
A new vector with new replacement values.
}
\description{
Replaces multiple values in a vector according to pairs provided.
}
\examples{
vvv = LETTERS[1:5]
fff = c('B', 'D')
ttt = c('XX', 'YY')
vvv = replace_values(vec=vvv, from=fff, to=ttt)
print(vvv)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
