% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublist.R
\name{sublist}
\alias{sublist}
\title{Vector to a nested list with a name}
\usage{
sublist(vect, subname = NULL)
}
\arguments{
\item{vect}{A vector, which you want to convert to a nested list.}

\item{subname}{The name of each element of the vector in each nested list.}
}
\value{
A nested list with its element being each element of the vector with its name as specified.
}
\description{
Convert each element of a vector to a nested list with a name.
}
\examples{
abc = LETTERS[1:3]
lst = sublist(vect=abc, subname='alphabet')
print(lst)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
