% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Value replacement in a dataframe}
\usage{
replace_values(
  dat,
  from = c("aː", "a", "iː", "ɪ", "uː", "ʊ"),
  to = c("a", "a", "i", "i", "u", "u"),
  clm = "Segment"
)
}
\arguments{
\item{dat}{A dataframe to which you want to execute replacements.}

\item{from}{A vector of strings which you want to replace for something else.}

\item{to}{A vector of strings with the same length as 'from'. Target strings specified by 'from' are replaced by the corresponding values specified by this variable ('to').}

\item{clm}{A string, which gives the name of the column whose values you want to replace.}
}
\value{
A new dataframe, whose column with the name specified by 'clm' now contains new replaced values.
}
\description{
Replaces multiple values in a dataframe according to pairs provided.
}
\examples{
ddd = data.frame('letter'=LETTERS[1:5], 'value'=seq(10,50,10))
fff = c('B', 'D')
ttt = c('XX', 'YY')
dat = replace_values(dat=ddd, from=fff, to=ttt, clm='letter')
print(dat)
}
\author{
Motoki Saito, \email{motoki.saito@uni-tuebingen.de}
}
\keyword{utilities}
